config: partner_gms: Remove hardcoded variant list
Instead of hardcoding the GMS variant, allow to set the makefiles
which should be included for WITH_GMS builds.

For example:
  Project FI
    - WITH_GMS_FI := true
    + GMS_MAKEFILE := fi.mk

  Go devices
    - WITH_GMS_GO := true
    + GMS_MAKEFILE := gms_go.mk
    + MAINLINE_MODULES_MAKEFILE := mainline_modules_low_ram.mk

  Mainline modules without updatable apex
    + MAINLINE_MODULES_MAKEFILE := mainline_modules_flatten_apex.mk

As you expect GMS to be configured the way you specified them, this
change also causes the build to error out, if a makefile can not be
included.

This excludes mainline modules to allow including GMS without using
mainline modules.

Change-Id: I0ab61449ed7cc60e72bf519f91c29712055f8a74